refactor(velite): extract prepare hook into testable phases
velite.config.ts:prepare was a 256-line arrow function with cyclomatic
24 / cognitive 44 / CRAP 600 (the largest hotspot in the repo by CRAP).
The complexity was cosmetic — the function did 5 distinct, sequential
phases that compose cleanly. As written they were impossible to unit-test
because the whole thing was a closure inside a `defineConfig` literal.

This change moves each phase to `src/lib/velite-prepare.ts` as a named,
typed, exported function:

1. filterDrafts          — production-only draft removal
2. validateCrossReferences — SCHEMA-04 / D-01..D-04 cross-ref integrity
3. buildAndWriteGraphs   — dependency graph SVG render + graph.json
4. extractGitHistory     — git-history.json
5. versionAndValidateArtifacts — SCHEMA-08 / D-05..D-07 hash gate +
   CalVer derivation + Zod shape validation
6. writeRegistryFiles    — public/r/ JSON files for CLI/HTTP consumption
   (relocated from velite.config.ts; was already extracted)

velite.config.ts:prepare is now a 22-line orchestrator. Behavior is
preserved exactly — verified by byte-identical .velite/artifacts.json
and .velite/multiArtifacts.json against main with the same input.

Adds tests/velite-prepare-cross-refs.test.ts — 8 unit tests covering
the three failure modes (invalid format / unknown collection / missing
target), the happy path, cross-collection refs (D-02), nested-slug
path-shaped refs, and accumulator-then-throw behavior (D-04). Runs in
0.24s vs the existing fixture-based integration test which spawns
velite per assertion.

Verified:
- pnpm typecheck — clean
- pnpm build — clean (.velite, public/pagefind, public/r/ all generated)
- pnpm test (phase27 cross-ref + dup-slug + voice + calver + perf-baseline
  + content-pipeline + new velite-prepare-cross-refs) — 31/31 pass

Fallow health: velite.config.ts:prepare critical CRAP 600 finding is gone.
Replaced by two moderate findings in velite-prepare.ts (CRAP 37 + None
for the now-unit-tested validateCrossReferences).
