+++ apps/blakepetersen.io/velite.config.ts
@@ -1,7 +1,6 @@
 // ABOUTME: Velite content pipeline configuration with typed schemas for all collections.
 // ABOUTME: Defines content and artifact collections, merges artifacts in prepare hook into artifacts.json.
 
-import { createHash } from 'node:crypto'
 import fs from 'node:fs'
 import path from 'node:path'
 import { defineCollection, defineConfig, s } from 'velite'
@@ -10,20 +9,14 @@ import rehypeShiki from '@shikijs/rehype'
 import { transformerMetaHighlight } from '@shikijs/transformers'
 import { terminalTheme } from './src/lib/shiki-theme'
 import {
-  buildGraph,
-  getLocalGraph,
-  computeLayout,
-  renderGraphSvg,
-} from './src/lib/graph'
-import type { ContentNode } from './src/lib/graph'
-import { getGitHistoryForFile } from './src/lib/git-history'
-import { deriveCalVer } from './src/lib/calver'
-import {
-  SlugSchema,
-  CalVerSchema,
-  ArtifactTypeSchema,
-  MergeStrategySchema,
-} from 'blink-registry'
+  filterDrafts,
+  validateCrossReferences,
+  buildAndWriteGraphs,
+  extractGitHistory,
+  versionAndValidateArtifacts,
+  writeRegistryFiles,
+  type DxData,
+} from './src/lib/velite-prepare'
 
 // DX collection names — single source of truth for cross-ref prefix validation.
 const DX_COLLECTIONS = ['skills', 'hooks', 'configs', 'guides'] as const
@@ -216,330 +209,29 @@ const config: any = defineConfig({
   },
   collections: { skills, hooks, configs, guides, posts, singleArtifacts, multiArtifacts },
   strict: true,
-  prepare: (data) => {
-    const isProduction = process.env.NODE_ENV === 'production'
-    if (isProduction) {
-      data.skills = data.skills.filter((s) => !s.draft)
-      data.hooks = data.hooks.filter((h) => !h.draft)
-      data.configs = data.configs.filter((c) => !c.draft)
-      data.guides = data.guides.filter((g) => !g.draft)
-      data.posts = data.posts.filter((p) => !p.draft)
+  prepare: (data: DxData) => {
+    if (process.env.NODE_ENV === 'production') {
+      filterDrafts(data)
     }
 
-    // Collect all DX content items (not posts) for dependency graph
-    const dxItems: ContentNode[] = [
-      ...data.skills,
-      ...data.hooks,
-      ...data.configs,
-      ...data.guides,
-    ].map((item) => ({
-      slug: item.slug,
-      title: item.title,
-      category: item.category,
-      dependencies: item.dependencies,
-    }))
+    validateCrossReferences(data)
 
-    // SCHEMA-04: cross-reference integrity check (D-01..D-04).
-    // Runs against post-draft-filter data so dev = prod for valid published
-    // content. Walks `dependencies` and `related` only — `decisions` is wrong
-    // shape (no slugs) per D-01. Cross-collection refs allowed per D-02.
-    // Format is `<collection>/<slug>` per D-03; bare slug may itself contain
-    // slashes (nested skills), so we compare against the path-shaped slug
-    // directly rather than splitting again. Accumulator-then-throw per D-04.
-    type DxCollectionName = 'skills' | 'hooks' | 'configs' | 'guides'
-    const collectionSlugs: Record<DxCollectionName, Set<string>> = {
-      skills: new Set(data.skills.map((i) => i.slug)),
-      hooks: new Set(data.hooks.map((i) => i.slug)),
-      configs: new Set(data.configs.map((i) => i.slug)),
-      guides: new Set(data.guides.map((i) => i.slug)),
-    }
-
-    const brokenRefs: string[] = []
-    const allDx: Array<{ slug: string; dependencies?: string[]; related?: string[] }> = [
-      ...data.skills,
-      ...data.hooks,
-      ...data.configs,
-      ...data.guides,
-    ]
-
-    for (const item of allDx) {
-      for (const field of ['dependencies', 'related'] as const) {
-        const refs = (item[field] ?? []) as string[]
-        for (const ref of refs) {
-          const slashIndex = ref.indexOf('/')
-          if (slashIndex <= 0 || slashIndex === ref.length - 1) {
-            brokenRefs.push(
-              `  ${item.slug} ${field}: '${ref}' — invalid format (expected '<collection>/<slug>')`,
-            )
-            continue
-          }
-          const coll = ref.slice(0, slashIndex)
-          if (!(coll in collectionSlugs)) {
-            brokenRefs.push(
-              `  ${item.slug} ${field}: '${ref}' — unknown collection '${coll}'`,
-            )
-            continue
-          }
-          // ref is the full path-shaped slug (e.g. 'skills/claude-code/writing-custom-skills');
-          // collection's Set stores the same path-shaped value, so compare directly.
-          if (!collectionSlugs[coll as DxCollectionName].has(ref)) {
-            brokenRefs.push(
-              `  ${item.slug} ${field}: '${ref}' — target not found in collection '${coll}'`,
-            )
-          }
-        }
-      }
-    }
-
-    if (brokenRefs.length > 0) {
-      throw new Error(
-        `Broken cross-references in DX content (${brokenRefs.length}):\n${brokenRefs.join('\n')}`,
-      )
-    }
-
-    const fullGraph = buildGraph(dxItems)
-    const fullLayout = computeLayout(fullGraph)
-    const fullGraphSvg = renderGraphSvg(fullLayout, { basePath: '/dx' })
-
-    const localGraphs: Record<string, string> = {}
-    for (const item of dxItems) {
-      const local = getLocalGraph(fullGraph, item.slug)
-      if (local.edges.length > 0) {
-        const localLayout = computeLayout(local)
-        localGraphs[item.slug] = renderGraphSvg(localLayout, {
-          currentSlug: item.slug,
-          basePath: '/dx',
-        })
-      }
-    }
-
-    const graphData = { fullGraphSvg, localGraphs }
     const outputDir = path.resolve(process.cwd(), '.velite')
     fs.mkdirSync(outputDir, { recursive: true })
-    fs.writeFileSync(
-      path.join(outputDir, 'graph.json'),
-      JSON.stringify(graphData, null, 2),
-    )
 
-    // Extract git history for all content items
+    buildAndWriteGraphs(data, outputDir)
+
     const contentDir = path.resolve(process.cwd(), 'content')
-    const allItems = [
-      ...data.skills,
-      ...data.hooks,
-      ...data.configs,
-      ...data.guides,
-      ...data.posts,
-    ]
+    extractGitHistory(data, contentDir, outputDir)
 
-    const gitHistory: Record<string, { lastModified: string; commitCount: number }> = {}
-    for (const item of allItems) {
-      const mdxPath = path.join(contentDir, `${item.slug}.mdx`)
-      gitHistory[item.slug] = getGitHistoryForFile(mdxPath)
-    }
-
-    fs.writeFileSync(
-      path.join(outputDir, 'git-history.json'),
-      JSON.stringify(gitHistory, null, 2),
-    )
-
-    // Merge single-file and multi-file artifacts into unified artifacts.json
-    function deriveArtifactSlug(velitePath: string): string {
-      const cleaned = velitePath
-        .replace(/\.artifact\/manifest$/, '')
-        .replace(/\.artifact$/, '')
-      // Extract just the filename portion to produce a valid slug (no slashes)
-      return cleaned.split('/').pop() || cleaned
-    }
-
-    // SCHEMA-08: load the git-tracked version manifest before deriveCalVer
-    // is ever called. The hash gate short-circuits version derivation when
-    // an artifact's distributed-payload bytes are unchanged (D-05/D-07),
-    // which is what preserves "prose-only edits don't bump CalVer" even
-    // though deriveCalVer is NOT pure-date (it shells out to git log).
-    // Manifest is git-tracked per D-06 so versions stay deterministic
-    // across machines after first capture. Single-process serial build
-    // assumption (Risk #3) — no atomic write needed for v1.4.
-    const versionManifestPath = path.resolve(contentDir, '.artifact-versions.json')
-    type VersionManifest = Record<string, { hash: string; version: string }>
-    const priorVersionManifest: VersionManifest = fs.existsSync(versionManifestPath)
-      ? JSON.parse(fs.readFileSync(versionManifestPath, 'utf-8'))
-      : {}
-    const updatedVersionManifest: VersionManifest = {}
-    const dateCounters = new Map<string, number>()
-
-    function sha256Hex(payload: string): string {
-      return createHash('sha256').update(payload).digest('hex')
-    }
-
-    const singles = (data.singleArtifacts ?? []).map((artifact) => {
-      const slug = deriveArtifactSlug(artifact.slug)
-      // D-05: hash the distributed-payload bytes only (single-file = body).
-      const hash = sha256Hex(artifact.body)
-      const prior = priorVersionManifest[slug]
-      let version: string
-      if (prior && prior.hash === hash) {
-        version = prior.version
-      } else {
-        const filePath = path.join(contentDir, `${artifact.slug}.md`)
-        version = deriveCalVer(filePath, dateCounters)
-      }
-      updatedVersionManifest[slug] = { hash, version }
-
-      return {
-        slug,
-        name: artifact.name,
-        type: artifact.type,
-        version,
-        description: artifact.description,
-        files: [
-          {
-            path: artifact.destination,
-            content: artifact.body,
-            merge: artifact.merge,
-          },
-        ],
-        devDependencies: artifact.devDependencies,
-      }
-    })
-
-    const multis = (data.multiArtifacts ?? []).map((artifact) => {
-      const slug = deriveArtifactSlug(artifact.slug)
-      // D-05: hash concatenated file.content in declared order (no separator —
-      // boundaries are immaterial to the consumer-facing payload).
-      const concatenated = artifact.files.map((f) => f.content).join('')
-      const hash = sha256Hex(concatenated)
-      const prior = priorVersionManifest[slug]
-      let version: string
-      if (prior && prior.hash === hash) {
-        version = prior.version
-      } else {
-        const manifestFilePath = path.join(contentDir, `${artifact.slug}.json`)
-        version = deriveCalVer(manifestFilePath, dateCounters)
-      }
-      updatedVersionManifest[slug] = { hash, version }
-
-      return {
-        slug,
-        name: artifact.name,
-        type: artifact.type,
-        version,
-        description: artifact.description,
-        files: artifact.files,
-        devDependencies: artifact.devDependencies,
-      }
-    })
-
-    const allArtifacts = [...singles, ...multis]
-
-    // Validate artifact shape at build time (fail-fast)
-    for (const artifact of allArtifacts) {
-      if (!SlugSchema.safeParse(artifact.slug).success) {
-        throw new Error(`Invalid artifact slug: "${artifact.slug}"`)
-      }
-      if (!CalVerSchema.safeParse(artifact.version).success) {
-        throw new Error(`Invalid artifact version: "${artifact.version}" for ${artifact.slug}`)
-      }
-      if (!ArtifactTypeSchema.safeParse(artifact.type).success) {
-        throw new Error(`Invalid artifact type: "${artifact.type}" for ${artifact.slug}`)
-      }
-      for (const file of artifact.files) {
-        if (!MergeStrategySchema.safeParse(file.merge).success) {
-          throw new Error(`Invalid merge strategy: "${file.merge}" in ${artifact.slug}`)
-        }
-        if (!file.content) {
-          throw new Error(`Empty file content for "${file.path}" in ${artifact.slug}`)
-        }
-      }
-    }
-
-    // SCHEMA-08: persist updated version manifest after artifact validation
-    // passes. Single-process serial build (Risk #3) — no atomic write needed.
-    // Sort top-level keys so consecutive runs are byte-identical (D-06: diffable in PRs).
-    const sortedVersionManifest = Object.fromEntries(
-      Object.entries(updatedVersionManifest).sort(([a], [b]) => (a < b ? -1 : a > b ? 1 : 0)),
-    )
-    fs.writeFileSync(
-      versionManifestPath,
-      JSON.stringify(sortedVersionManifest, null, 2) + '\n',
-    )
-
+    const allArtifacts = versionAndValidateArtifacts(data, contentDir)
     fs.writeFileSync(
       path.join(outputDir, 'artifacts.json'),
       JSON.stringify(allArtifacts, null, 2),
     )
 
-    // Generate static registry JSON files for CLI and HTTP consumption
     writeRegistryFiles(allArtifacts)
   },
 })
-
-function writeRegistryFiles(
-  allArtifacts: Array<{
-    slug: string
-    name: string
-    type: string
-    version: string
-    description: string
-    files: Array<{ path: string; content: string; merge: string }>
-    devDependencies?: Record<string, string>
-    dependencies?: string[]
-  }>,
-) {
-  const baseUrl =
-    process.env.NEXT_PUBLIC_SITE_URL || 'https://blakepetersen.io'
-  const registryDir = path.resolve(process.cwd(), 'public', 'r')
-
-  // Clean stale files
-  fs.rmSync(registryDir, { recursive: true, force: true })
-  fs.mkdirSync(registryDir, { recursive: true })
-
-  // Build index with url per item
-  const items = allArtifacts.map((artifact) => ({
-    slug: artifact.slug,
-    name: artifact.name,
-    type: artifact.type,
-    version: artifact.version,
-    description: artifact.description,
-    url: `${baseUrl}/${artifact.type}s/${artifact.slug}`,
-  }))
-
-  // Use max CalVer version as generatedAt (avoids noisy git diffs from wall-clock time)
-  const maxVersion =
-    allArtifacts.length > 0
-      ? allArtifacts
-          .map((a) => a.version)
-          .sort()
-          .pop()!
-      : new Date().toISOString()
-
-  // CalVer versions aren't ISO datetimes, so convert to a stable ISO timestamp
-  // Format: YYYY.MM.DD.N -> YYYY-MM-DDT00:00:00Z
-  const generatedAt = maxVersion.match(/^\d{4}\.\d{2}\.\d{2}\.\d+$/)
-    ? `${maxVersion.split('.').slice(0, 3).join('-')}T00:00:00Z`
-    : maxVersion
-
-  const index = { items, generatedAt }
-
-  fs.writeFileSync(
-    path.join(registryDir, 'index.json'),
-    JSON.stringify(index, null, 2),
-  )
-
-  // Write per-artifact detail files
-  for (const artifact of allArtifacts) {
-    const typeDir = path.join(registryDir, artifact.type)
-    fs.mkdirSync(typeDir, { recursive: true })
-
-    const detail = {
-      ...artifact,
-      url: `${baseUrl}/${artifact.type}s/${artifact.slug}`,
-    }
-
-    fs.writeFileSync(
-      path.join(typeDir, `${artifact.slug}.json`),
-      JSON.stringify(detail, null, 2),
-    )
-  }
-}
 
 export default config
